+++ notification-handler/spec/factories/notifications.rb
@@ -3,6 +3,5 @@
 FactoryBot.define do
   factory :notification do
     target { build_stubbed :user }
-    group { nil }
   end
 end