+++ notification-handler/spec/support/rails_app/config/initializers/notification_handler.rb
@@ -9,7 +9,13 @@ NotificationHandler.configure do |config|
   # that don't necessarily have a common class.
   # Learn more: https://github.com/jonhue/notifications-rails/tree/master/notification-handler#groups
   config.define_group :subscribers, -> { User.where(subscriber: true) }
-  config.define_group :subscribed,  lambda { |subscribed|
+  config.define_group :subscribed, lambda { |subscribed|
     User.where(subscriber: subscribed)
   }
+  config.define_group :subscribed_and_starts_with,
+                      lambda { |subscribed, starts_with|
+                        User.where(subscriber: subscribed)
+                            .where('name LIKE :prefix',
+                                   prefix: "#{starts_with}%")
+                      }
 end